Immagine spezzata correttamente
Ad ogni stream viene assegnata una sottoregione dell'immagine, che viene
trasferita su host in maniera asincrona.
Una volta su host, queste sottoregioni vengono renderizzate sulla
finestra, andando a rimappare le coordinate edlle singole regioni
nell'immagine definitiva.

Fare tutti quei conti mi ha fatto venire un gran mal di testa.

+++ header/Fract.h
@@ -19,6 +19,9 @@
 #include <Pixel.h>
 #include <common.h>
 
+#include <ctime>
+#include <sdf_util.hpp>
+
 #define WIDTH 512.0
 #define HEIGHT 512.0
 #define MAX_STEPS 128
@@ -26,11 +29,11 @@
 #define BLOCK_DIM_X 8
 #define BLOCK_DIM_Y 8
 #define NUM_STREAMS 16
-#define PIXEL_PER_STREAM_X (WIDTH / 4)
-#define PIXEL_PER_STREAM_Y (HEIGHT / 4)
-#define PIXEL_PER_STREAM ((WIDTH / 4)*(HEIGHT / 4))
+#define PIXEL_PER_STREAM_X (int)(WIDTH / 4)
+#define PIXEL_PER_STREAM_Y (int)(HEIGHT / 4)
+#define PIXEL_PER_STREAM (int)((WIDTH / 4)*(HEIGHT / 4))
 
-typedef pixel pixelRegionForStream[(int)PIXEL_PER_STREAM];
+typedef pixel pixelRegionForStream[PIXEL_PER_STREAM];
 
 class Fract
 {
